Don't use an iterator in the network pool
For whatever reason, using a persistent iterator in an AutoNetwork::Pool
within the official Vagrant 1.2.2 ruby on OSX Mountain Lion causes
FiberErrors, as follows:

    /Users/zalgo/src/vagrants/vagrant-auto_network/lib/auto_network/pool.rb:22:in `next': fiber called across threads (FiberError)
        from /Users/zalgo/src/vagrants/vagrant-auto_network/lib/auto_network/pool.rb:22:in `next'
        from /Users/zalgo/src/vagrants/vagrant-auto_network/lib/auto_network/mixin.rb:11:in `auto_network!'
        from /Users/zalgo/src/vagrants/hiera_mysql/Vagrantfile:48:in `block (2 levels) in <top (required)>'
    [stack trace continues]

It's unclear what is triggering this, but it looks to be an ugly
intersection of how iterators handle state intersecting with some
internal threading. This commit removes the iterator and instead mutates
an array of available addresses as they are fetched.