Extend acceptance testing to status and reload
Ensure that the `status` and `reload` commands continue to work when
auto_network is loaded. Vagrant 1.6.0 introduced some internal changes that
broke `reload`.

Also partially reverts cddc317 by adding back a `vagrant destroy` run in an
`after` block. This is redundant, but ensures everything is cleand up if the
acceptance test fails part of the way through.