Set acceptance IP ranges through environment
This paves the way for running multiple providers under acceptance tests. The
obvious candidate is VMware Fusion, but there are still some kinks to hammer
out.

This patch passes in the IP Address range used during acceptance through the
environment. It also does away with the shared contexts used to set these
variables as they were broken. Instead, variables are set in the top-level
config.