Better plugin loading for acceptance tests
By loading the `gemspec` inside of the Bundler `:plugins` group, we tap into
the Bundler-based plugin loader shipped with v1.5.4. This is better than using
`require` at the top of the Vagrantfile as it is closer to the way things
execute in reality.

This probably means we won't be able to support running acceptance tests
against 1.4.x and earlier. However, that isn't a great loss as the currently
pined version of `vagrant-spec` won't work with 1.4.x.