Store global pool manager instance in settings
After 1.6, we can no longer rely on the Vagrant Environment instance to persist
plugin-specific state. This means that values set during environment
initialization hooks will not survive to be present for the actions that need
them.

This patch moves the storage of state to the AutoNetwork module its self.

(Also, integration tests. Yey!)

Ref: mitchellh/vagrant#3660