Update LoadPool Action for Vagrant 1.6
As of 1.6, actions called from the `:environment_load` hook no longer have
access to some information. The workaround is to use the `env[:env]` key.

The situation is complicated by the fact that actions called outside of
`:environment_load` may be passed an environment that does not inclue
`env[:env]`. The fix for these situations is to only run the `LoadPool` action
during the `:environment_load` hook.

Ref: mitchellh/vagrant#3701