Ignore unconfigured machines from Vagrant env
If a VM definition is removed from a Vagrantfile but the VM is not destroyed,
an id will be left behind in the `.vagrant` directory and will show up in
`env.active_machines`. Attempting to fetch a machine using this id will raise
`Vagrant::Errors::MachineNotFound`.

This patch traps `MachineNotFound` errors and ignores the associated ids when
composing lists of active Vagrant VMs.

Copied from adrienthebo/vagrant-hosts#34.