Handle pre-allocated IPs in the Request action
This patch fixes a bug that occurred when a newly created VM already had an IP
allocated in the AutoNetwork pool. The `FilterNetworks` action would not fire
because the VM was not created. The `Request` action would not fire because an
IP had already been allocated.

This patch extends the `Request` action such that it ensures no unfiltered
AutoNetwork networks remain on a VM in addition to allocating an IP if
necessary.

Fixes GH-23.